Fixed ability to find existing email when registering and added ability to send listing data in basic questions seciton.
..
Add: Extra details fields to be saved in sharedpreferences in LocationFragment so that it can be saved to the database.
Add: add insertListingData() in DatabaseInterface to insert listing data.
Add: PublishListingDataRequest as JSON body so that data can be sent with insertListingData() in DatabaseInterface.
Add: user_id in SessionManager so that the database can track down how many listings the user created.
Add:idListing to receive id of listing after listing is published. Listing id is used to track down images included in the listing.
...