Able to book a schedule for a listing.
..
Add: Instead of replacing the fragments when a layoutis clicked (like layoutAmenities), hide and add the fragment with fragment tags instead. When back button is called, pop the backstack
Add: methods in the DatabaseInterface to get/post/update a booking schedule for the listing.
Add: POJO to get the JSON object of booking schedules.
Add: BookingSendSMSFragment to send SMS to the host when the user books.
Add: Feature in AvailabilityFragment when user can only book dates before a booked date.
...