Added search-bar and it's animation to enlarge/minimize in HomeFragment. Able to have search listings through location filters and guest filters. Location filters (LocationFragment) will filter the listing by a specific country/street/state/city. Guest filter (SearchGuestFragment) can filter the number of total guest (including infants and children) and filter the permission on pets to stay.
..
Add: parameteres in multipleListingsData and totalListingsData for the listing filters. If the param is empty the API will assign it as a widlcard (%).
Add: In GuestFragment, removed any other characters  except the guest number. This is so that when inserted to the database, it would be a string number which can easily be converted to an int.
..